major/major/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.5.2
pandas==2.1.4
numpy==1.26.2
email-validator==2.0.0
python-multipart==0.0.6
Werkzeug==2.3.7
